pyproject.toml
[project]
name = "gmail-cleaner"
version = "1.0.0"
description = "A free, privacy-focused tool to bulk unsubscribe, delete emails, and mark as read"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Gururagavendra" }
]
keywords = ["gmail", "email", "unsubscribe", "cleanup", "privacy"]

dependencies = [
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-api-python-client>=2.100.0",
    "fastapi>=0.122.0",
    "uvicorn>=0.38.0",
    "pydantic-settings>=2.11.0",
]

[project.urls]
Homepage = "https://github.com/Gururagavendra/gmail-cleaner"
Repository = "https://github.com/Gururagavendra/gmail-cleaner"

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []
